update Customer web API

+++ src/main/java/com/bit/final_project/dto/entityDto/DoorQuotationDto.java
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Setter
 public class DoorQuotationDto {
     private String id;
-    private Long height;
-    private Long width;
+    private Double height;
+    private Double width;
     private String color;
     private DesignDto design;
 }
